Move plugin PHP logic into namespace file
